MINOR: Move `RaftRequestHandler` to `tools` package
